split out validation logic
